requirements.txt
discord.py
requests
aiohttp
asyncio
